Module Cohttp_async.Response

include Cohttp.S.Response with type t = Cohttp.Response.t

type t = Cohttp.Response.t = {

encoding : Cohttp.Transfer.encoding;
Transfer encoding of this HTTP response

headers : Cohttp.Header.t;
response HTTP headers

version : Cohttp.Code.version;
(** HTTP version, usually 1.1 *)

status : Cohttp.Code.status_code;
HTTP status code of the response

flush : bool;

}

include Ppx_sexp_conv_lib.Sexpable.S with type Response.t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val encoding : t -> Cohttp.Transfer.encoding

val headers : t -> Cohttp.Header.t

val version : t -> Cohttp.Code.version

val status : t -> Cohttp.Code.status_code

val flush : t -> bool

val compare : t -> t -> int

val make : ?⁠version:Cohttp.Code.version -> ?⁠status:Cohttp.Code.status_code -> ?⁠flush:bool -> ?⁠encoding:Cohttp.Transfer.encoding -> ?⁠headers:Cohttp.Header.t -> unit -> t

The response creates by make ~encoding ~headers () has an encoding value determined from the content of headers or if no proper header is present, using the value of encoding. Checked headers are "content-length", "content-range" and "transfer-encoding". The default value of encoding is chunked.